+++ R/spex.R
@@ -73,3 +73,51 @@ spex.Extent <- function(x, crs, ...) {
   crs(p) <- crs
   spex(p, ...)
 }
+
+#' Extent of simple features
+#' 
+#' This is the simplest of the missing "raster support" for the sf package, 
+#' here using the xmin, xmax, ymin, ymax convention used by raster rather than 
+#' the transpose version favoured in sp and sf. 
+#' @param x object with an extent
+#' @param ... unused 
+#' @name extent
+#' @aliases Extent
+#' @importFrom raster extent
+extent_sf <- function(x, ...) {
+  raster::extent(attr(x[[attr(x, "sf_column")]], "bbox")[c(1, 3, 2, 4)])
+}
+setOldClass("sf")
+setMethod(f = "extent", signature = "sf", definition = extent_sf)
+
+#' @export
+#' @name spex
+spex.sf <- function(x, crs, ...) {
+  spex(extent(x), attr(x[[attr(x, "sf_column")]], "crs")$proj4string)
+}
+
+#' Axis ranges from extent
+#' 
+#' Functions `xlim` and `ylim` return the two-value counterparts of an extent. 
+#' 
+#' Any projection metadata is dropped since this is a one-dimensional entity. 
+#' @param x any object with an extent understood by `spex`
+#' @param ... reserved for future methods
+#'
+xlim <- function(x, ...) UseMethod("xlim")
+#' @export
+#' @name spex
+xlim.default <- function(x, ...) {
+  spx <- spex(x)
+  c(raster::xmin(spx), raster::xmax(spx))
+}
+#' @export
+#' @name spex
+ylim <- function(x, ...) UseMethod("ylim")
+#' @export
+#' @name spex
+ylim.default <- function(x, ...) {
+  spx <- spex(x)
+  c(raster::ymin(spx), raster::ymax(spx))
+}
+