+++ data-raw/psf.R
@@ -1,0 +1,4 @@
+
+library(scsf)
+psf <- minimal_mesh
+devtools::use_data(psf)