+++ tests/testthat/test-spex-extent.R
@@ -10,3 +10,13 @@ test_that("by id is not implemented", {
 })
 
 
+
+test_that("extent of anything", {
+  library(raster)
+  raster::extent(psf)
+  expect_equal(raster::xmin(spex(psf, "+")), 0)
+  expect_equal(raster::xmax(spex(psf, "+")), 1.23)
+  
+  graphics.off()
+  expect_equal(raster::extent(spex()) , raster::extent(0, 1, 0, 1))
+})